Refactor "Add Folder" button into a separate file
